+++ api/V0AdjustClusterResources.go
@@ -3,7 +3,6 @@ package api
 import (
 	"net/http"
 	"strconv"
-	"strings"
 
 	logrus "github.com/AVENTER-UG/mesos-m3s/logger"
 	"github.com/gorilla/mux"
@@ -15,58 +14,75 @@ import (
 func (e *API) V0AdjustClusterResources(w http.ResponseWriter, r *http.Request) {
 	logrus.WithField("func", "api.V0AdjustClusterResources").Debug("Call")
 
-	vars := mux.Vars(r)
-
-	pathFragments := strings.Split(r.URL.Path, "/")
-
-	component := pathFragments[4]
-	resource := pathFragments[5]
-
 	if !e.CheckAuth(r, w) {
 		w.WriteHeader(http.StatusUnauthorized)
 		return
 	}
 
+	vars := mux.Vars(r)
 	if vars == nil {
+		w.WriteHeader(http.StatusNotAcceptable)
+		return
+	}
+
+	component := vars["component"]
+	resource := vars["value"]
+	value, err := strconv.Atoi(vars["value"])
+	if component == "" || resource == "" || err != nil {
+		logrus.WithField("func", "api.V0AdjustClusterResources").Error("Error: Required Fields/Parameters are empty or wrong: ", err.Error())
 		w.WriteHeader(http.StatusNotAcceptable)
 		return
 	}
 
 	d := e.ErrorMessage(0, "V0AdjustClusterResources", "ok")
 
-	if vars["value"] != "" {
-		value, err := strconv.Atoi(vars["value"])
-		if err != nil {
-			logrus.WithField("func", "api.V0AdjustClusterResources").Error("Error: ", err.Error())
-			w.WriteHeader(http.StatusNotAcceptable)
-			return
+	switch component {
+	case "ds":
+		if resource == "cpus" {
+			e.Config.DSCPU = float64(value)
+		} else if resource == "cpusLimit" {
+			e.Config.DSCPULimit = float64(value)
+		} else if resource == "memory" {
+			e.Config.DSMEM = float64(value)
+		} else if resource == "memoryLimit" {
+			e.Config.DSMEMLimit = float64(value)
+		} else if resource == "disk" {
+			e.Config.DSDISK = float64(value)
+		} else if resource == "diskLimit" {
+			e.Config.DSDISKLimit = float64(value)
 		}
-
-		switch component {
-		case "ds":
-			if resource == "cpus" {
-				e.Config.DSCPU = float64(value)
-			} else if resource == "memory" {
-				e.Config.DSMEM = float64(value)
-			}
-		case "server":
-			if resource == "cpus" {
-				e.Config.K3SServerCPU = float64(value)
-			} else if resource == "memory" {
-				e.Config.K3SServerMEM = float64(value)
-			}
-		case "agent":
-			if resource == "cpus" {
-				e.Config.K3SAgentCPU = float64(value)
-			} else if resource == "memory" {
-				e.Config.K3SAgentMEM = float64(value)
-			}
+	case "server":
+		if resource == "cpus" {
+			e.Config.K3SServerCPU = float64(value)
+		} else if resource == "cpusLimit" {
+			e.Config.K3SServerCPULimit = float64(value)
+		} else if resource == "memory" {
+			e.Config.K3SServerMEM = float64(value)
+		} else if resource == "memoryLimit" {
+			e.Config.K3SServerMEMLimit = float64(value)
+		} else if resource == "disk" {
+			e.Config.K3SServerDISK = float64(value)
+		} else if resource == "diskLimit" {
+			e.Config.K3SServerDISKLimit = float64(value)
 		}
-
-		e.Redis.SaveConfig(*e.Config)
-		d = []byte("Resources Updated")
-
+	case "agent":
+		if resource == "cpus" {
+			e.Config.K3SAgentCPU = float64(value)
+		} else if resource == "cpusLimit" {
+			e.Config.K3SAgentCPULimit = float64(value)
+		} else if resource == "memory" {
+			e.Config.K3SAgentMEM = float64(value)
+		} else if resource == "memoryLimit" {
+			e.Config.K3SAgentMEMLimit = float64(value)
+		} else if resource == "disk" {
+			e.Config.K3SAgentDISK = float64(value)
+		} else if resource == "diskLimit" {
+			e.Config.K3SAgentDISKLimit = float64(value)
+		}
 	}
+
+	e.Redis.SaveConfig(*e.Config)
+	d = []byte("Resources Updated")
 
 	w.Header().Set("Content-Type", "application/json; charset=utf-8")
 	w.Header().Set("Api-Service", "v0")