+++ shell.nix
@@ -4,11 +4,12 @@ stdenv.mkDerivation {
 name = "go-env";
 
 buildInputs = [
-    go
-    syft
-    grype
-    docker
-    trivy
+		go
+		syft
+		grype
+		docker
+		docker-credential-helpers
+		trivy
 ];
 
 SOURCE_DATE_EPOCH = 315532800;
@@ -16,8 +17,10 @@ PROJDIR = "${toString ./.}";
 S_NETWORK="host";
 
 shellHook = ''
-    export LD_LIBRARY_PATH="${pkgs.stdenv.cc.cc.lib}/lib"
-    export PATH=/tmp/bin:$PATH
-    mkdir /tmp/bin
-    '';
+		export LD_LIBRARY_PATH="${pkgs.stdenv.cc.cc.lib}/lib"
+		export PATH=/tmp/bin:$PATH
+		export GOTMPDIR=/tmp
+		export TMPDIR=/tmp
+		mkdir /tmp/bin
+		'';
 }